src/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pydantic==2.6.1
pydantic_core==2.16.2
pytest==8.0.0
pytest-asyncio==0.23.5
python-dotenv==1.0.1
sniffio==1.3.0
SQLAlchemy==2.0.26
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.27.1
